Changed ghost_pattern type to Uint16 because MusSeqPattern::pattern is Uint16

+++ trunk/src/mused.h
@@ -94,7 +94,7 @@ typedef struct
 		pattern_position, sequence_position, pattern_horiz_position, sequence_horiz_position,
 		program_position, current_program_step,
 		edit_reverb_param;
-	int *ghost_pattern[MUS_MAX_CHANNELS];
+	Uint16 *ghost_pattern[MUS_MAX_CHANNELS];
 	int current_sequencetrack;
 	Uint16 time_signature;
 	Clipboard cp;
